feat(app): add layouts and middleware
