+++ src/app/layout.tsx
@@ -6,9 +6,15 @@ import { Toaster } from '@/components/ui/toaster';
 import NextTopLoader from 'nextjs-toploader';
 import RenewSession from '@/app/(auth)/components/molecules/RenewSession';
 import { clsx } from 'clsx';
+import AppProvider from '@/store/app-provider';
+import { cookies } from 'next/headers';
+import userApiRequest from '@/api/user';
+import { FullUser } from '@/schema/user.schema';
+import { QueryProvider } from '@/queries/QueryProvider';
 
 const roboto = Roboto({
   weight: ['400', '700', '500', '900'],
+
   style: ['normal', 'italic'],
   subsets: ['latin'],
   display: 'swap',
@@ -21,11 +27,19 @@ export const metadata: Metadata = {
   metadataBase: new URL('https://next-learn-dashboard.vercel.sh'),
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const cookieStore = cookies();
+  const sessionToken = cookieStore.get('sessionToken');
+  let user: FullUser | null = null;
+  if (sessionToken) {
+    const res = await userApiRequest.me(sessionToken.value);
+    user = res.data;
+  }
+
   return (
     <html lang="en" suppressHydrationWarning>
       <body className={clsx(roboto.className, 'overflow-hidden')}>
@@ -36,8 +50,12 @@ export default function RootLayout({
           enableSystem
           disableTransitionOnChange
         >
-          {children}
-          <RenewSession />
+          <QueryProvider>
+            <AppProvider inititalSessionToken={sessionToken?.value} user={user}>
+              {children}
+              <RenewSession />
+            </AppProvider>
+          </QueryProvider>
         </ThemeProvider>
         <Toaster />
       </body>
